🏗️ Architect: Refactor analytics-debug.ts into modules
Split the monolithic `src/ui/analytics-debug.ts` (1023 lines) into smaller maintainable modules:
- `src/ui/analytics-debug-ui.ts`: Styles, DOM creation, and visual rendering updates.
- `src/ui/analytics-debug-handlers.ts`: Overlay state, control logic, and console setup.
- `src/ui/analytics-debug.ts`: A barrel file that transparently re-exports the modules to keep backward compatibility with downstream imports.

Updated `REFACTORING_PLAN_REMAINING.md` progress to 70% and synced `plan.md` with implementation details, proposing `src/audio/audio-system.ts` as the next step.
