+++ REFACTORING_PLAN_REMAINING.md
@@ -2,7 +2,7 @@
 
 This document outlines the remaining 5 files (5,275 lines) that need to be refactored to break them into smaller, more manageable modules. This is a continuation of the large file refactoring task.
 
-**Status**: 60% Complete (6 of 10 files refactored)
+**Status**: 70% Complete (7 of 10 files refactored)
 
 ## Remaining Files Summary
 
@@ -12,7 +12,7 @@ This document outlines the remaining 5 files (5,275 lines) that need to be refac
 | src/systems/region-manager.ts | 1064 | 3 files | ✅ DONE |
 | src/ui/loading-screen.ts | 1063 | 3 files | ✅ DONE |
 | src/audio/audio-system.ts | 1047 | 3 files | ⏳ TODO |
-| src/ui/analytics-debug.ts | 1023 | 3 files | ⏳ TODO |
+| src/ui/analytics-debug.ts | 1023 | 3 files | ✅ DONE |
 | **TOTAL** | **5,275** | **16 files** | **⏳ 0%** |
 
 ## Detailed Refactoring Plans
